Rename `MON_ID` to `MON_OT_ID` (#1114)
* Rename MON_ID to MON_TRAINER_ID

The constant name MON_ID implies it relates to the ID number of the
Pokémon itself. However, it actually refers to the Trainer ID number of
the Pokémon's Original Trainer. Renaming it to MON_TRAINER_ID makes it
substantially more clear what its purpose is.

The original definition of MON_ID is preserved in macros/legacy.asm

* Remove MON_ID from macros/legacy.asm

Co-authored-by: Sylvie <35663410+Rangi42@users.noreply.github.com>

* Apply rename to missed file

* Rename to `MON_OT_ID`

---------

Co-authored-by: SnorlaxMonster <snorlaxmonster@users.noreply.github.com>
Co-authored-by: Sylvie <35663410+Rangi42@users.noreply.github.com>
Co-authored-by: vulcandth <vulcandth@gmail.com>